@@ -6,9 +6,9 @@ namespace DynamicSample
 {
     public class WroxDynamicObject : DynamicObject
     {
-        private Dictionary<string, object> _dynamicData = new Dictionary<string, object>();
+        private readonly Dictionary<string, object?> _dynamicData = new Dictionary<string, object?>();
 
-        public override bool TryGetMember(GetMemberBinder binder, out object result)
+        public override bool TryGetMember(GetMemberBinder binder, out object? result)
         {
             bool success = false;
             result = null;
@@ -25,17 +25,25 @@ namespace DynamicSample
             return success;
         }
 
-        public override bool TrySetMember(SetMemberBinder binder, object value)
+        public override bool TrySetMember(SetMemberBinder binder, object? value)
         {
             _dynamicData[binder.Name] = value;
             return true;
         }
 
-        public override bool TryInvokeMember(InvokeMemberBinder binder, object[] args, out object result)
+        public override bool TryInvokeMember(InvokeMemberBinder binder, object?[]? args, out object? result)
         {
-            dynamic method = _dynamicData[binder.Name];
-            result = method((DateTime)args[0]);
-            return result != null;
+            result = null;
+            dynamic? method = _dynamicData[binder.Name];
+            if (method == null || args == null || args?.Length < 1 || args?[0] == null)
+            {
+                return false;
+            }
+            else
+            {
+                result = method((DateTime)args[0]);
+                return true;
+            }
         }
     }
 }
